requirements.txt
python3-memcached
tornado
jinja2
git+https://github.com/fsufitch/YAUL.git

